(fix) No longer avoid repeating v8bot and gbot2
At some point in time, purr was in the same rooms as other bots that
would evaluate javascript code, and so there was a special case to avoid
also evaluating javascript when these bots were in the room. Discord
doesn't have these bots, so we can safely remove this code rather than
porting it to Discord.