(new) Factoids now support rich user data
Previously, the factoids implementation used user nicknames to track who
modified a factoid, but Discord tracks id, username, discriminator and,
in channels, nicknames. We now save and store this information in the
factoids database and surface the nickname and, if different from the
user's current nickname, their nickname at the time of modification.