(new) Vendor µpaws.js
I *believe* that Heroku will handle submodules poorly. There are a few
other alternatives to this approach: namely, upgrading to Paws.js and
installing with npm and git+https; adding a package.json to µpaws.js and
installing with npm and git+https; and removing the paws functionality
completely. I eschewed these strategies for the following reasons:

- Paws.js is incomplete and currently doesn't actualyl execute code.
  µpaws is also incomplete, but "less incomplete".
- µpaws.js is currently unmaintained, so it feels wrong to add commits
  to it now.
- The current functionality *does* do things, and removing it completely
  would make me feel a little sad.