(new) Add RedisSaver class
The RedisSaver class exposes roughly the same interface as JSONSaver.
This minimizes the amount of refactoring necessary to switch from
JSONSaver. It *does* mean that this will technically scale poorly, so we
may need to refactor this more heavily at a later date.

The RedisSaver also emits events on load and on flush, which will be
useful for migration purposes.