(re) Use Knex instead of raw pg client
This doesn't actually work on Heroku, and here's why:

Heroku requires that you connect to their supplied databases using SSL
on and allowing self-signed certificates. This totally works with the
"pg" module directly - set the "ssl" option to { rejectUnauthorized:
false } - but the Knex module wraps this connection and doesn't support
bleeding the option through.